# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Kiokudo Core
  version: 0.2.0-staging
  description: >-
    Staging-only migration API. Review/card endpoints require a configured staging DB
    and private web-to-core service authentication. Production cutover is NOT approved.
paths:
  /api/v1/health:
    get:
      summary: Public minimal process liveness
      responses:
        '200': { description: Process reachable, not database readiness }
  /api/v1/status:
    get:
      summary: Authenticated migration status
      security: [{ ServiceBearer: [] }]
      responses:
        '200': { description: Migration gates and staging DB availability }
        '401': { description: Unauthorized }
  /api/v1/cards:
    get:
      summary: Read cards and deck summaries from staging database
      security: [{ ServiceBearer: [] }]
      parameters:
        - { in: query, name: deck, schema: { type: string } }
        - { in: query, name: search, schema: { type: string } }
        - { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 2000 } }
      responses:
        '200': { description: Legacy-compatible cards/deck summary response }
        '401': { description: Unauthorized }
        '503': { description: Staging DB not configured }
  /api/v1/reviews:
    post:
      summary: Atomically update the card state and append an immutable FSRS review event
      security: [{ ServiceBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cardId, rating]
              properties:
                eventId: { type: string, description: Stable idempotency key; strongly recommended }
                cardId: { type: string }
                rating: { oneOf: [{ type: string, enum: [Again, Hard, Good, Easy] }, { type: integer, minimum: 1, maximum: 4 }] }
                reviewedAt: { type: string, format: date-time }
                responseTimeMs: { type: integer }
      responses:
        '200': { description: Review applied or idempotent duplicate }
        '400': { description: Validation failed }
        '404': { description: Card not found }
        '409': { description: Event ID collision }
        '500': { description: Atomic transaction failed }
        '503': { description: Staging DB not configured }
  /api/v1/reviews/batch:
    post:
      summary: Offline event replay in ascending reviewedAt order; single DB transaction
      security: [{ ServiceBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reviews]
              properties:
                reviews:
                  type: array
                  minItems: 1
                  maxItems: 200
                  items:
                    type: object
                    required: [eventId, cardId, rating, reviewedAt]
                    properties:
                      eventId: { type: string }
                      cardId: { type: string }
                      rating: { type: string }
                      reviewedAt: { type: string, format: date-time }
      responses:
        '200': { description: Per-event applied/duplicate/rejected results }
        '400': { description: Invalid batch shape }
        '500': { description: Unexpected failure; batch must rollback }
        '503': { description: Staging DB not configured }
components:
  securitySchemes:
    ServiceBearer: { type: http, scheme: bearer, bearerFormat: opaque }